requirements.txt
torch>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
netCDF4>=1.6.0
xarray>=2023.1.0
tqdm>=4.65.0
pyyaml>=6.0
tensorboard>=2.13.0
